# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: E-Sync Telemetries Engine API
  description: E-Sync Telemetries Engine API documentation. This API provides access to GPS tracking, OBD data, DTC codes, fuel consumption, and other telemetry features.
  contact:
    name: E-Sync Support
    url: https://www.encipher.co.tz/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0.0'
servers:
  - url: /v1
    description: E-Sync Telemetries Engine API (v1)
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /positions:
    get:
      summary: Fetches a list of Positions
      description: >-
        We strongly recommend using [Traccar WebSocket
        API](https://www.traccar.org/traccar-api/) instead of periodically
        polling positions endpoint. Without any params, it returns a list of
        last known positions for all the user's Devices. _from_ and _to_ fields
        are not required with _id_.
      tags:
        - Positions
      parameters:
        - name: deviceId
          in: query
          description: >-
            _deviceId_ is optional, but requires the _from_ and _to_ parameters
            when used
          schema:
            type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          schema:
            type: string
            format: date-time
        - name: id
          in: query
          description: >-
            To fetch one or more positions. Multiple params can be passed like
            `id=31&id=42`
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
            text/csv:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
            application/gpx+xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
    delete:
      summary: Deletes all the Positions of a device in the time span specified
      description: ''
      tags:
        - Positions
      parameters:
        - name: deviceId
          in: query
          description: ''
          schema:
            type: integer
          required: true
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          schema:
            type: string
            format: date-time
          required: true
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          schema:
            type: string
            format: date-time
          required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content: {}
  /positions/{id}:
    delete:
      summary: Delete a Position
      tags:
        - Positions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
          content: {}
        '404':
          description: Not Found
          content: {}
...
  /v1/dtc:
    get:
      summary: Fetch a list of DTC codes
      description: Retrieve Diagnostic Trouble Codes (DTC) for devices
      tags:
        - DTC Codes
      parameters:
        - name: deviceId
          in: query
          description: Filter by device ID
          schema:
            type: integer
        - name: status
          in: query
          description: Filter by status (active, pending, permanent, cleared)
          schema:
            type: string
        - name: from
          in: query
          description: Start time in ISO 8601 format
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: End time in ISO 8601 format
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DtcCode'
  /v1/dtc/{id}:
    get:
      summary: Get DTC code by ID
      tags:
        - DTC Codes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtcCode'
        '404':
          description: Not Found
          content: {}
  /v1/dtc/device/{deviceId}:
    get:
      summary: Get DTC codes for a device
      tags:
        - DTC Codes
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          description: Filter by status (active, pending, permanent, cleared)
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DtcCode'
  /v1/dtc/device/{deviceId}/active:
    get:
      summary: Get active DTC codes for a device
      tags:
        - DTC Codes
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DtcCode'
  /v1/dtc/{id}/clear:
    put:
      summary: Clear a DTC code
      description: Mark a DTC code as cleared
      tags:
        - DTC Codes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtcCode'
        '404':
          description: Not Found
          content: {}
...
components:
  schemas:
    Position:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique position record identifier
        deviceId:
          type: integer
          format: int64
          description: Identifier of the device that reported this position
        protocol:
          type: string
          description: Device protocol name that produced the message
        deviceTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        fixTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        serverTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        valid:
          type: boolean
          description: Indicates whether the position was marked as valid by the device
        latitude:
          type: number
          description: Latitude in decimal degrees
        longitude:
          type: number
          description: Longitude in decimal degrees
        altitude:
          type: number
          description: Altitude above sea level in meters
        speed:
          type: number
          description: in knots
        course:
          type: number
          description: Heading in degrees (0-360) where 0 is true north
        address:
          type: string
          description: Resolved reverse-geocoded address if available
        accuracy:
          type: number
          description: Estimated positional accuracy in meters when provided
        network:
          type: object
          description: Network metadata (e.g. cell or Wi‑Fi data) supplied by the device
          properties: {}
        geofenceIds:
          type: array
          items:
            type: integer
          description: List of geofence ids applicable to this position
        attributes:
          type: object
          description: Custom key-value attributes sent by the device or enrichments
          properties: {}
...
    DtcCode:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique DTC code record identifier
        deviceId:
          type: integer
          format: int64
          description: Identifier of the device
        positionId:
          type: integer
          format: int64
          description: Identifier of the associated position
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the DTC code was recorded
        code:
          type: string
          description: DTC code (e.g., "P0301")
        codeType:
          type: string
          description: Code type (P, B, C, U)
          nullable: true
        description:
          type: string
          description: Human-readable description of the DTC code
          nullable: true
        status:
          type: string
          description: Status (active, pending, permanent, cleared)
        freezeFrameData:
          type: string
          description: Freeze frame data in JSON format
          nullable: true
        firstOccurred:
          type: string
          format: date-time
          description: When the DTC code first occurred
          nullable: true
        lastOccurred:
          type: string
          format: date-time
          description: When the DTC code last occurred
          nullable: true
        clearedAt:
          type: string
          format: date-time
          description: When the DTC code was cleared
          nullable: true
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer